Fix divide by -1.
This problem affects all platforms. It was introduced by commit
ccae333844c7aad0934f13f7698894c883a6b561 (Load SIMD constants with IR_FLOAD
from GG_State)

The error is that lj_ir_ksimd is used to construct the constant in op2 of
IR_NEG, but this function actually generates an IR_FLOAD of the constant
instead. In doing so, it overwrites fins (ie J->fold.ins). The fix is to
directly reference the required constant in simplify_nummuldiv_k.